smart pricing working

+++ Unique_for_printing.xaml.cs
@@ -20,7 +20,7 @@ namespace kiosk_snapprint
         public string ColorMode { get; }
         public List<int> SelectedPages { get; }
         public int CopyCount { get; }
-        public int TotalPrice { get; }
+        public double TotalPrice { get; }
 
         // Add properties for overlay visibility
         public Visibility OverlayVisibility
@@ -29,7 +29,7 @@ namespace kiosk_snapprint
             set { LoadingOverlay.Visibility = value; }
         }
 
-        public Unique_for_printing(byte[] fileBytes, string fileName, string pageSize, string colorMode, List<int> selectedPages, int copyCount, int totalPrice)
+        public Unique_for_printing(byte[] fileBytes, string fileName, string pageSize, string colorMode, List<int> selectedPages, int copyCount, double totalPrice)
         {
             InitializeComponent();
             // Set the values to class properties
